overlay/master: missed leaveDotGit
When adding the `postFetch` for creating the necessary `.git_head` and
`.git_time` values, setting `leaveDotGit` was missed. This is necessary to allow
the fetch to stay as a Git repository. Note that `.git` is removed as the final
step – so the directory will no longer be a Git repository.